Work around std::set iterator issue with const_cast
C++11 enforces const set iterators, to avoid invalidating the
containers' 'sorted' invariant.  If an object stored in the set is
changed in such a way that this cannot happen (i.e., the parts used to
sort them remain unchanged), then the change should be allowed.

The workarounds available are (1) making everything that isn't used to
sort the objects mutable, (2) to use const_cast to change the object,
and (3) to erase the object and re-insert a changed copy.
(1) completely subverts const correctness for every other use of the
class, opening up larger risks.
(2) is considered bad form, and allows sets to become unsorted if used
carelessly.
(3) adds unnecessary overhead, but is the only conceptually safe option.
I have opted for (2), and must be as careful to avoid invalidating sets'
sorted nature as I have been before switching to VS2012.  If anything,
having to go through an explicit const_cast will remind me to be careful
of what I am changing.