Clear _items early in loadData
Loading objects and NPCs may result in new items being specified and
pointed to (materials, loot, etc.), and so clearing _items was moved
ahead of checking those files.

There is still room for this problem to reappear, if loadData() is
called once with object data specifying items, and again with item data.