Change collision chunks when entities move
This fixes a bug whereby an entity that move sufficiently far away from
its starting location becomes invisible to the collision system.  As a
consequence, other entities feel free to occupy the same space, and when
the invisible entity tries to move it fails a valid-location assertion.

That assertion has now been removed for performance reasons.