Add user/container/slot info to users' item instances
This should be enough information for the ServerItem::Instance to report updates to their owners.  For now it excludes items in object containers, but this will need to change if any part of item state can change while that item is in a container; e.g., if you can repair an item while it's in a pot, or if quantity is rolled into the Instance class.