Remove virtual function call in base-class c'tor
Ideally this wouldn't be called in any c'tor, but for now this works.  If Object or NPC are ever subclassed then there will be problems.

The specific problem in this case was that when an object was created, Entity::onSetType() was called but not Object::onSetType, resulting in merchant slots not being created.  An attempt to populate them when a static object was created led to a server crash.