Make spawned NPC teleport home if its target becomes unreachable
Before this fix, they kept trying to pathfind over and over again.  The problem manifested as NPCs crowding around the dungeon gate after players had left.

Part of the problem was that the MOB ownership type was added, but not properly excluded from being treated like a pet following its owner when its target disappeared.

The other issue was that when it came time to find a path, the NPC didn't know whether a path had already been attempted with the current constraints, allowing infinite tries.